Add GCS support for searchable snapshots
