+++ plugins/repository-s3/src/test/java/org/elasticsearch/repositories/s3/S3BlobContainerRetriesTests.java
@@ -21,18 +21,13 @@ package org.elasticsearch.repositories.s3;
 import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.internal.MD5DigestCalculatingInputStream;
 import com.amazonaws.util.Base16;
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpServer;
-import org.apache.http.ConnectionClosedException;
 import org.apache.http.HttpStatus;
-import org.apache.http.NoHttpResponseException;
 import org.elasticsearch.cluster.metadata.RepositoryMetadata;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.SuppressForbidden;
 import org.elasticsearch.common.blobstore.BlobContainer;
 import org.elasticsearch.common.blobstore.BlobPath;
 import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.common.io.Streams;
 import org.elasticsearch.common.lucene.store.ByteArrayIndexInput;
 import org.elasticsearch.common.lucene.store.InputStreamIndexInput;
@@ -44,27 +39,18 @@ import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.util.concurrent.CountDown;
 import org.elasticsearch.core.internal.io.IOUtils;
-import org.elasticsearch.mocksocket.MockHttpServer;
-import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.repositories.blobstore.AbstractBlobContainerRetriesTestCase;
 import org.junit.After;
 import org.junit.Before;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.SocketTimeoutException;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.NoSuchFileException;
-import java.util.Arrays;
 import java.util.Locale;
 import java.util.Objects;
-import java.util.Optional;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static org.elasticsearch.repositories.s3.S3ClientSettings.DISABLE_CHUNKED_ENCODING;
 import static org.elasticsearch.repositories.s3.S3ClientSettings.ENDPOINT_SETTING;
@@ -72,41 +58,47 @@ import static org.elasticsearch.repositories.s3.S3ClientSettings.MAX_RETRIES_SET
 import static org.elasticsearch.repositories.s3.S3ClientSettings.READ_TIMEOUT_SETTING;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.either;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.lessThan;
-import static org.hamcrest.Matchers.lessThanOrEqualTo;
 
 /**
  * This class tests how a {@link S3BlobContainer} and its underlying AWS S3 client are retrying requests when reading or writing blobs.
  */
 @SuppressForbidden(reason = "use a http server")
-public class S3BlobContainerRetriesTests extends ESTestCase {
+public class S3BlobContainerRetriesTests extends AbstractBlobContainerRetriesTestCase {
 
-    private static final long MAX_RANGE_VAL = Long.MAX_VALUE - 1;
-
-    private HttpServer httpServer;
     private S3Service service;
 
     @Before
     public void setUp() throws Exception {
         service = new S3Service();
-        httpServer = MockHttpServer.createHttp(new InetSocketAddress(InetAddress.getLoopbackAddress(), 0), 0);
-        httpServer.start();
         super.setUp();
     }
 
     @After
     public void tearDown() throws Exception {
         IOUtils.close(service);
-        httpServer.stop(0);
         super.tearDown();
     }
 
-    private BlobContainer createBlobContainer(final @Nullable Integer maxRetries,
+    @Override
+    protected String downloadStorageEndpoint(String blob) {
+        return "/bucket/" + blob;
+    }
+
+    @Override
+    protected String bytesContentType() {
+        return "text/plain; charset=utf-8";
+    }
+
+    @Override
+    protected Class<? extends Exception> unresponsiveExceptionType() {
+        return SdkClientException.class;
+    }
+
+    @Override
+    protected BlobContainer createBlobContainer(final @Nullable Integer maxRetries,
                                               final @Nullable TimeValue readTimeout,
                                               final @Nullable Boolean disableChunkedEncoding,
                                               final @Nullable ByteSizeValue bufferSize) {
@@ -144,210 +136,6 @@ public class S3BlobContainerRetriesTests extends ESTestCase {
             repositoryMetadata));
     }
 
-    public void testReadNonexistentBlobThrowsNoSuchFileException() {
-        final BlobContainer blobContainer = createBlobContainer(between(1, 5), null, null, null);
-        final long position = randomLongBetween(0, MAX_RANGE_VAL);
-        final int length = randomIntBetween(0, Math.toIntExact(Math.min(Integer.MAX_VALUE, MAX_RANGE_VAL - position)));
-        final Exception exception = expectThrows(NoSuchFileException.class,
-            () -> {
-                if (randomBoolean()) {
-                    blobContainer.readBlob("read_nonexistent_blob");
-                } else {
-                    blobContainer.readBlob("read_nonexistent_blob", 0, 1);
-                }
-            });
-        assertThat(exception.getMessage().toLowerCase(Locale.ROOT), containsString("blob object [read_nonexistent_blob] not found"));
-        assertThat(expectThrows(NoSuchFileException.class, () -> blobContainer.readBlob("read_nonexistent_blob", position, length))
-            .getMessage().toLowerCase(Locale.ROOT), containsString("blob object [read_nonexistent_blob] not found"));
-    }
-
-    public void testReadBlobWithRetries() throws Exception {
-        final int maxRetries = randomInt(5);
-        final CountDown countDown = new CountDown(maxRetries + 1);
-
-        final byte[] bytes = randomBlobContent();
-        httpServer.createContext("/bucket/read_blob_max_retries", exchange -> {
-            Streams.readFully(exchange.getRequestBody());
-            if (countDown.countDown()) {
-                final int rangeStart = getRangeStart(exchange);
-                assertThat(rangeStart, lessThan(bytes.length));
-                assertEquals(Optional.empty(), getRangeEnd(exchange));
-                exchange.getResponseHeaders().add("Content-Type", "text/plain; charset=utf-8");
-                exchange.sendResponseHeaders(HttpStatus.SC_OK, bytes.length - rangeStart);
-                exchange.getResponseBody().write(bytes, rangeStart, bytes.length - rangeStart);
-                exchange.close();
-                return;
-            }
-            if (randomBoolean()) {
-                exchange.sendResponseHeaders(randomFrom(HttpStatus.SC_INTERNAL_SERVER_ERROR, HttpStatus.SC_BAD_GATEWAY,
-                                                        HttpStatus.SC_SERVICE_UNAVAILABLE, HttpStatus.SC_GATEWAY_TIMEOUT), -1);
-            } else if (randomBoolean()) {
-                sendIncompleteContent(exchange, bytes);
-            }
-            if (randomBoolean()) {
-                exchange.close();
-            }
-        });
-
-        final TimeValue readTimeout = TimeValue.timeValueSeconds(between(1, 3));
-        final BlobContainer blobContainer = createBlobContainer(maxRetries, readTimeout, null, null);
-        try (InputStream inputStream = blobContainer.readBlob("read_blob_max_retries")) {
-            final int readLimit;
-            final InputStream wrappedStream;
-            if (randomBoolean()) {
-                // read stream only partly
-                readLimit = randomIntBetween(0, bytes.length);
-                wrappedStream = Streams.limitStream(inputStream, readLimit);
-            } else {
-                readLimit = bytes.length;
-                wrappedStream = inputStream;
-            }
-            final byte[] bytesRead = BytesReference.toBytes(Streams.readFully(wrappedStream));
-            logger.info("maxRetries={}, readLimit={}, byteSize={}, bytesRead={}",
-                maxRetries, readLimit, bytes.length, bytesRead.length);
-            assertArrayEquals(Arrays.copyOfRange(bytes, 0, readLimit), bytesRead);
-            if (readLimit < bytes.length) {
-                // we might have completed things based on an incomplete response, and we're happy with that
-            } else {
-                assertTrue(countDown.isCountedDown());
-            }
-        }
-    }
-
-    public void testReadRangeBlobWithRetries() throws Exception {
-        final int maxRetries = randomInt(5);
-        final CountDown countDown = new CountDown(maxRetries + 1);
-
-        final byte[] bytes = randomBlobContent();
-        httpServer.createContext("/bucket/read_range_blob_max_retries", exchange -> {
-            Streams.readFully(exchange.getRequestBody());
-            if (countDown.countDown()) {
-                final int rangeStart = getRangeStart(exchange);
-                assertThat(rangeStart, lessThan(bytes.length));
-                assertTrue(getRangeEnd(exchange).isPresent());
-                final int rangeEnd = getRangeEnd(exchange).get();
-                assertThat(rangeEnd, greaterThanOrEqualTo(rangeStart));
-                // adapt range end to be compliant to https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35
-                final int effectiveRangeEnd = Math.min(bytes.length - 1, rangeEnd);
-                final int length = (effectiveRangeEnd - rangeStart) + 1;
-                exchange.getResponseHeaders().add("Content-Type", "text/plain; charset=utf-8");
-                exchange.sendResponseHeaders(HttpStatus.SC_OK, length);
-                exchange.getResponseBody().write(bytes, rangeStart, length);
-                exchange.close();
-                return;
-            }
-            if (randomBoolean()) {
-                exchange.sendResponseHeaders(randomFrom(HttpStatus.SC_INTERNAL_SERVER_ERROR, HttpStatus.SC_BAD_GATEWAY,
-                    HttpStatus.SC_SERVICE_UNAVAILABLE, HttpStatus.SC_GATEWAY_TIMEOUT), -1);
-            } else if (randomBoolean()) {
-                sendIncompleteContent(exchange, bytes);
-            }
-            if (randomBoolean()) {
-                exchange.close();
-            }
-        });
-
-        final TimeValue readTimeout = TimeValue.timeValueMillis(between(100, 500));
-        final BlobContainer blobContainer = createBlobContainer(maxRetries, readTimeout, null, null);
-        final int position = randomIntBetween(0, bytes.length - 1);
-        final int length = randomIntBetween(0, randomBoolean() ? bytes.length : Integer.MAX_VALUE);
-        try (InputStream inputStream = blobContainer.readBlob("read_range_blob_max_retries", position, length)) {
-            final int readLimit;
-            final InputStream wrappedStream;
-            if (randomBoolean()) {
-                // read stream only partly
-                readLimit = randomIntBetween(0, length);
-                wrappedStream = Streams.limitStream(inputStream, readLimit);
-            } else {
-                readLimit = length;
-                wrappedStream = inputStream;
-            }
-            final byte[] bytesRead = BytesReference.toBytes(Streams.readFully(wrappedStream));
-            logger.info("maxRetries={}, position={}, length={}, readLimit={}, byteSize={}, bytesRead={}",
-                maxRetries, position, length, readLimit, bytes.length, bytesRead.length);
-            assertArrayEquals(Arrays.copyOfRange(bytes, position, Math.min(bytes.length, position + readLimit)), bytesRead);
-            if (readLimit == 0 || (readLimit < length && readLimit == bytesRead.length)) {
-                // we might have completed things based on an incomplete response, and we're happy with that
-            } else {
-                assertTrue(countDown.isCountedDown());
-            }
-        }
-    }
-
-    public void testReadBlobWithReadTimeouts() {
-        final int maxRetries = randomInt(5);
-        final TimeValue readTimeout = TimeValue.timeValueMillis(between(100, 200));
-        final BlobContainer blobContainer = createBlobContainer(maxRetries, readTimeout, null, null);
-
-        // HTTP server does not send a response
-        httpServer.createContext("/bucket/read_blob_unresponsive", exchange -> {});
-
-        Exception exception = expectThrows(SdkClientException.class, () -> blobContainer.readBlob("read_blob_unresponsive"));
-        assertThat(exception.getMessage().toLowerCase(Locale.ROOT), containsString("read timed out"));
-        assertThat(exception.getCause(), instanceOf(SocketTimeoutException.class));
-
-        // HTTP server sends a partial response
-        final byte[] bytes = randomBlobContent();
-        httpServer.createContext("/bucket/read_blob_incomplete", exchange -> sendIncompleteContent(exchange, bytes));
-
-        final int position = randomIntBetween(0, bytes.length - 1);
-        final int length = randomIntBetween(1, randomBoolean() ? bytes.length : Integer.MAX_VALUE);
-        exception = expectThrows(IOException.class, () -> {
-            try (InputStream stream = randomBoolean() ?
-                    blobContainer.readBlob("read_blob_incomplete") :
-                    blobContainer.readBlob("read_blob_incomplete", position, length)) {
-                Streams.readFully(stream);
-            }
-        });
-        assertThat(exception, either(instanceOf(SocketTimeoutException.class)).or(instanceOf(ConnectionClosedException.class)));
-        assertThat(exception.getMessage().toLowerCase(Locale.ROOT), either(containsString("read timed out")).or(
-            containsString("premature end of chunk coded message body: closing chunk expected")));
-        assertThat(exception.getSuppressed().length, equalTo(maxRetries));
-    }
-
-    public void testReadBlobWithNoHttpResponse() {
-        final BlobContainer blobContainer = createBlobContainer(randomInt(5), null, null, null);
-
-        // HTTP server closes connection immediately
-        httpServer.createContext("/bucket/read_blob_no_response", HttpExchange::close);
-
-        Exception exception = expectThrows(SdkClientException.class,
-            () -> {
-                if (randomBoolean()) {
-                    blobContainer.readBlob("read_blob_no_response");
-                } else {
-                    blobContainer.readBlob("read_blob_no_response", 0, 1);
-                }
-            });
-        assertThat(exception.getMessage().toLowerCase(Locale.ROOT), containsString("the target server failed to respond"));
-        assertThat(exception.getCause(), instanceOf(NoHttpResponseException.class));
-        assertThat(exception.getSuppressed().length, equalTo(0));
-    }
-
-    public void testReadBlobWithPrematureConnectionClose() {
-        final int maxRetries = randomInt(20);
-        final BlobContainer blobContainer = createBlobContainer(maxRetries, null, null, null);
-
-        // HTTP server sends a partial response
-        final byte[] bytes = randomBlobContent();
-        httpServer.createContext("/bucket/read_blob_incomplete", exchange -> {
-            sendIncompleteContent(exchange, bytes);
-            exchange.close();
-        });
-
-        final Exception exception = expectThrows(ConnectionClosedException.class, () -> {
-            try (InputStream stream = randomBoolean() ?
-                    blobContainer.readBlob("read_blob_incomplete", 0, 1):
-                    blobContainer.readBlob("read_blob_incomplete")) {
-                Streams.readFully(stream);
-            }
-        });
-        assertThat(exception.getMessage().toLowerCase(Locale.ROOT),
-            either(containsString("premature end of chunk coded message body: closing chunk expected"))
-                .or(containsString("premature end of content-length delimited message body")));
-        assertThat(exception.getSuppressed().length, equalTo(Math.min(S3RetryingInputStream.MAX_SUPPRESSED_EXCEPTIONS, maxRetries)));
-    }
-
     public void testWriteBlobWithRetries() throws Exception {
         final int maxRetries = randomInt(5);
         final CountDown countDown = new CountDown(maxRetries + 1);
@@ -372,7 +160,7 @@ public class S3BlobContainerRetriesTests extends ESTestCase {
                     } else {
                         Streams.readFully(exchange.getRequestBody());
                         exchange.sendResponseHeaders(randomFrom(HttpStatus.SC_INTERNAL_SERVER_ERROR, HttpStatus.SC_BAD_GATEWAY,
-                                                                HttpStatus.SC_SERVICE_UNAVAILABLE, HttpStatus.SC_GATEWAY_TIMEOUT), -1);
+                            HttpStatus.SC_SERVICE_UNAVAILABLE, HttpStatus.SC_GATEWAY_TIMEOUT), -1);
                     }
                 }
                 exchange.close();
@@ -503,141 +291,5 @@ public class S3BlobContainerRetriesTests extends ESTestCase {
         assertThat(countDownInitiate.isCountedDown(), is(true));
         assertThat(countDownUploads.get(), equalTo(0));
         assertThat(countDownComplete.isCountedDown(), is(true));
-    }
-
-    private static byte[] randomBlobContent() {
-        return randomByteArrayOfLength(randomIntBetween(1, frequently() ? 512 : 1 << 20)); // rarely up to 1mb
-    }
-
-    private static final Pattern RANGE_PATTERN = Pattern.compile("^bytes=([0-9]+)-([0-9]+)$");
-
-    private static Tuple<Long, Long> getRange(HttpExchange exchange) {
-        final String rangeHeader = exchange.getRequestHeaders().getFirst("Range");
-        if (rangeHeader == null) {
-            return Tuple.tuple(0L, MAX_RANGE_VAL);
-        }
-
-        final Matcher matcher = RANGE_PATTERN.matcher(rangeHeader);
-        assertTrue(rangeHeader + " matches expected pattern", matcher.matches());
-        long rangeStart = Long.parseLong(matcher.group(1));
-        long rangeEnd = Long.parseLong(matcher.group(2));
-        assertThat(rangeStart, lessThanOrEqualTo(rangeEnd));
-        return Tuple.tuple(rangeStart, rangeEnd);
-    }
-
-    private static int getRangeStart(HttpExchange exchange) {
-        return Math.toIntExact(getRange(exchange).v1());
-    }
-
-    private static Optional<Integer> getRangeEnd(HttpExchange exchange) {
-        final long rangeEnd = getRange(exchange).v2();
-        if (rangeEnd == MAX_RANGE_VAL) {
-            return Optional.empty();
-        }
-        return Optional.of(Math.toIntExact(rangeEnd));
-    }
-
-    private static void sendIncompleteContent(HttpExchange exchange, byte[] bytes) throws IOException {
-        final int rangeStart = getRangeStart(exchange);
-        assertThat(rangeStart, lessThan(bytes.length));
-        final Optional<Integer> rangeEnd = getRangeEnd(exchange);
-        final int length;
-        if (rangeEnd.isPresent()) {
-            // adapt range end to be compliant to https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35
-            final int effectiveRangeEnd = Math.min(rangeEnd.get(), bytes.length - 1);
-            length = effectiveRangeEnd - rangeStart + 1;
-        } else {
-            length = bytes.length - rangeStart;
-        }
-        exchange.getResponseHeaders().add("Content-Type", "text/plain; charset=utf-8");
-        exchange.sendResponseHeaders(HttpStatus.SC_OK, length);
-        final int bytesToSend = randomIntBetween(0, length - 1);
-        if (bytesToSend > 0) {
-            exchange.getResponseBody().write(bytes, rangeStart, bytesToSend);
-        }
-        if (randomBoolean()) {
-            exchange.getResponseBody().flush();
-        }
-    }
-
-    /**
-     * A resettable InputStream that only serves zeros.
-     **/
-    private static class ZeroInputStream extends InputStream {
-
-        private final AtomicBoolean closed = new AtomicBoolean(false);
-        private final long length;
-        private final AtomicLong reads;
-        private volatile long mark;
-
-        private ZeroInputStream(final long length) {
-            this.length = length;
-            this.reads = new AtomicLong(0);
-            this.mark = -1;
-        }
-
-        @Override
-        public int read() throws IOException {
-            ensureOpen();
-            return (reads.incrementAndGet() <= length) ? 0 : -1;
-        }
-
-        @Override
-        public int read(byte[] b, int off, int len) throws IOException {
-            ensureOpen();
-            if (len == 0) {
-                return 0;
-            }
-
-            final int available = available();
-            if (available == 0) {
-                return -1;
-            }
-
-            final int toCopy = Math.min(len, available);
-            Arrays.fill(b, off, off + toCopy, (byte) 0);
-            reads.addAndGet(toCopy);
-            return toCopy;
-        }
-
-        @Override
-        public boolean markSupported() {
-            return true;
-        }
-
-        @Override
-        public synchronized void mark(int readlimit) {
-            mark = reads.get();
-        }
-
-        @Override
-        public synchronized void reset() throws IOException {
-            ensureOpen();
-            reads.set(mark);
-        }
-
-        @Override
-        public int available() throws IOException {
-            ensureOpen();
-            if (reads.get() >= length) {
-                return 0;
-            }
-            try {
-                return Math.toIntExact(length - reads.get());
-            } catch (ArithmeticException e) {
-                return Integer.MAX_VALUE;
-            }
-        }
-
-        @Override
-        public void close() {
-            closed.set(true);
-        }
-
-        private void ensureOpen() throws IOException {
-            if (closed.get()) {
-                throw new IOException("Stream closed");
-            }
-        }
     }
 }
