+++ x-pack/plugin/searchable-snapshots/qa/gcs/build.gradle
@@ -7,7 +7,7 @@
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -16,10 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-
-import org.elasticsearch.gradle.MavenFilteringHack
 import org.elasticsearch.gradle.info.BuildParams
+import org.elasticsearch.gradle.MavenFilteringHack
 
 import java.nio.file.Files
 import java.security.KeyPair
@@ -29,21 +27,22 @@ import static org.elasticsearch.gradle.PropertyNormalization.IGNORE_VALUE
 
 apply plugin: 'elasticsearch.standalone-rest-test'
 apply plugin: 'elasticsearch.rest-test'
-apply plugin: 'elasticsearch.test.fixtures'
 
-// TODO think about flattening qa:google-cloud-storage project into parent
+final Project fixture = project(':test:fixtures:gcs-fixture')
+final Project repositoryPlugin = project(':plugins:repository-gcs')
+
 dependencies {
-  testCompile project(path: ':plugins:repository-gcs')
+  testCompile project(path: xpackModule('searchable-snapshots'), configuration: 'testArtifacts')
+  testCompile repositoryPlugin
 }
 
 restResources {
   restApi {
-    includeCore '_common', 'snapshot','indices', 'index', 'bulk', 'count'
+    includeCore 'indices', 'search', 'bulk', 'snapshot', 'nodes', '_common'
+    includeXpack 'searchable_snapshots'
   }
 }
 
-testFixtures.useFixture(':test:fixtures:gcs-fixture', 'gcs-fixture')
-testFixtures.useFixture(':test:fixtures:gcs-fixture', 'gcs-fixture-third-party')
 boolean useFixture = false
 
 String gcsServiceAccount = System.getenv("google_storage_service_account")
@@ -66,13 +65,6 @@ def encodedCredentials = {
   Base64.encoder.encodeToString(Files.readAllBytes(serviceAccountFile.toPath()))
 }
 
-def fixtureAddress = { fixture ->
-  assert useFixture: 'closure should not be used without a fixture'
-  int ephemeralPort = project(':test:fixtures:gcs-fixture').postProcessFixture.ext."test.fixtures.${fixture}.tcp.80"
-  assert ephemeralPort > 0
-  'http://127.0.0.1:' + ephemeralPort
-}
-
 /** A service account file that points to the Google Cloud Storage service emulated by the fixture **/
 task createServiceAccountFile() {
   doLast {
@@ -93,33 +85,12 @@ task createServiceAccountFile() {
   }
 }
 
-task thirdPartyTest(type: Test) {
-  if (useFixture) {
-    thirdPartyTest.dependsOn createServiceAccountFile
-    nonInputProperties.systemProperty 'test.google.endpoint', "${-> fixtureAddress('gcs-fixture-third-party')}"
-    nonInputProperties.systemProperty 'test.google.tokenURI', "${-> fixtureAddress('gcs-fixture-third-party')}/o/oauth2/token"
-
-    gradle.taskGraph.whenReady {
-      if (it.hasTask(gcsThirdPartyTests)) {
-        throw new IllegalStateException("Tried to run third party tests but not all of the necessary environment variables " +
-          "'google_storage_service_account', 'google_storage_bucket', 'google_storage_base_path' are set.")
-      }
-    }
-  }
-
-  include '**/GoogleCloudStorageThirdPartyTests.class'
-  systemProperty 'tests.security.manager', false
-  systemProperty 'test.google.bucket', gcsBucket
-  nonInputProperties.systemProperty 'test.google.base', gcsBasePath + "_third_party_tests_" + BuildParams.testSeed
-  nonInputProperties.systemProperty 'test.google.account', "${-> encodedCredentials.call()}"
+def fixtureAddress = { f ->
+  assert useFixture: 'closure should not be used without a fixture'
+  int ephemeralPort = project(':test:fixtures:gcs-fixture').postProcessFixture.ext."test.fixtures.${f}.tcp.80"
+  assert ephemeralPort > 0
+  'http://127.0.0.1:' + ephemeralPort
 }
-
-task gcsThirdPartyTests {
-  dependsOn check
-}
-
-integTest.mustRunAfter(thirdPartyTest)
-check.dependsOn thirdPartyTest
 
 Map<String, Object> expansions = [
   'bucket'   : gcsBucket,
@@ -131,21 +102,37 @@ processTestResources {
   MavenFilteringHack.filter(it, expansions)
 }
 
+if (useFixture) {
+  apply plugin: 'elasticsearch.test.fixtures'
+  testFixtures.useFixture(fixture.path, 'gcs-fixture-other')
+}
+
 integTest {
-  dependsOn project(':plugins:repository-gcs').bundlePlugin
+  dependsOn repositoryPlugin.bundlePlugin
+  runner {
+    systemProperty 'test.gcs.bucket', gcsBucket
+    systemProperty 'test.gcs.base_path', gcsBasePath + "/searchable_snapshots_tests"
+  }
 }
 
 testClusters.integTest {
-  plugin project(':plugins:repository-gcs').bundlePlugin.archiveFile
+  testDistribution = 'DEFAULT'
+  plugin repositoryPlugin.bundlePlugin.archiveFile
 
-  keystore 'gcs.client.integration_test.credentials_file', serviceAccountFile, IGNORE_VALUE
+  if (BuildParams.isSnapshotBuild() == false) {
+    systemProperty 'es.searchable_snapshots_feature_enabled', 'true'
+  }
 
+  keystore 'gcs.client.searchable_snapshots.credentials_file', serviceAccountFile, IGNORE_VALUE
   if (useFixture) {
     tasks.integTest.dependsOn createServiceAccountFile
     /* Use a closure on the string to delay evaluation until tests are executed */
-    setting 'gcs.client.integration_test.endpoint', { "${-> fixtureAddress('gcs-fixture')}" }, IGNORE_VALUE
-    setting 'gcs.client.integration_test.token_uri', { "${-> fixtureAddress('gcs-fixture')}/o/oauth2/token" }, IGNORE_VALUE
+    setting 'gcs.client.searchable_snapshots.endpoint', { "${-> fixtureAddress('gcs-fixture-other')}" }, IGNORE_VALUE
+    setting 'gcs.client.searchable_snapshots.token_uri', { "${-> fixtureAddress('gcs-fixture-other')}/o/oauth2/token" }, IGNORE_VALUE
   } else {
-    println "Using an external service to test the repository-gcs plugin"
+    println "Using an external service to test " + project.name
   }
+
+  setting 'xpack.license.self_generated.type', 'trial'
 }
+
